+++ .travis_script.bash
@@ -92,12 +92,9 @@ pony-stable-build-debs(){
   fi
 
   build_deb xenial
-  build_deb artful
   build_deb bionic
-  build_deb trusty
   build_deb buster
   build_deb stretch
-  build_deb jessie
 
   ls -la
   set +x
@@ -154,10 +151,7 @@ pony-stable-kickoff-copr-ppa(){
   sudo chmod 400 ~/.ssh/config
 
   build_and_submit_deb_src xenial
-  build_and_submit_deb_src artful
   build_and_submit_deb_src bionic
-  build_and_submit_deb_src cosmic
-  build_and_submit_deb_src trusty
 
   # COPR for fedora/centos/suse
   echo "Kicking off pony-stable packaging for COPR..."